Refactoring dashboard - create separate services
